Separate tracers' tests
