+++ google/tracer_test.go
@@ -1,0 +1,19 @@
+package google_test
+
+import (
+	"database/sql"
+
+	"github.com/go-sql-driver/mysql"
+	"github.com/luna-duclos/instrumentedsql"
+	"github.com/luna-duclos/instrumentedsql/google"
+)
+
+// WrapDriverGoogle demonstrates how to call wrapDriver and register a new driver.
+// This example uses MySQL and google tracing to illustrate this
+func ExampleWrapDriver_google() {
+	sql.Register("instrumented-mysql", instrumentedsql.WrapDriver(mysql.MySQLDriver{}, instrumentedsql.WithTracer(google.NewTracer(false))))
+	db, err := sql.Open("instrumented-mysql", "connString")
+
+	// Proceed to handle connection errors and use the database as usual
+	_, _ = db, err
+}
