+++ sql_example_test.go
@@ -5,44 +5,10 @@ import (
 	"database/sql"
 	"log"
 
-	"github.com/go-sql-driver/mysql"
-	"github.com/luna-duclos/instrumentedsql/opencensus"
 	"github.com/mattn/go-sqlite3"
 
 	"github.com/luna-duclos/instrumentedsql"
-	"github.com/luna-duclos/instrumentedsql/google"
-	"github.com/luna-duclos/instrumentedsql/opentracing"
 )
-
-// WrapDriverGoogle demonstrates how to call wrapDriver and register a new driver.
-// This example uses MySQL and google tracing to illustrate this
-func ExampleWrapDriver_google() {
-	sql.Register("instrumented-mysql", instrumentedsql.WrapDriver(mysql.MySQLDriver{}, instrumentedsql.WithTracer(google.NewTracer(false))))
-	db, err := sql.Open("instrumented-mysql", "connString")
-
-	// Proceed to handle connection errors and use the database as usual
-	_, _ = db, err
-}
-
-// WrapDriverOpentracing demonstrates how to call wrapDriver and register a new driver.
-// This example uses MySQL and opentracing to illustrate this
-func ExampleWrapDriver_opentracing() {
-	sql.Register("instrumented-mysql", instrumentedsql.WrapDriver(mysql.MySQLDriver{}, instrumentedsql.WithTracer(opentracing.NewTracer(false))))
-	db, err := sql.Open("instrumented-mysql", "connString")
-
-	// Proceed to handle connection errors and use the database as usual
-	_, _ = db, err
-}
-
-// WrapDriverOpencensus demonstrates how to call wrapDriver and register a new driver.
-// This example uses MySQL and opencensus to illustrate this
-func ExampleWrapDriver_opencensus() {
-	sql.Register("instrumented-mysql", instrumentedsql.WrapDriver(mysql.MySQLDriver{}, instrumentedsql.WithTracer(opencensus.NewTracer(false))))
-	db, err := sql.Open("instrumented-mysql", "connString")
-
-	// Proceed to handle connection errors and use the database as usual
-	_, _ = db, err
-}
 
 // WrapDriverJustLogging demonstrates how to call wrapDriver and register a new driver.
 // This example uses sqlite, and does not trace, but merely logs all calls
